move html to Welcome.tsx + fix p > ul
