+++ apps/hyparquet-demo/src/main.tsx
@@ -6,9 +6,6 @@ import './index.css'
 const app = document.getElementById('app')
 if (!app) throw new Error('missing app element')
 
-const params = new URLSearchParams(location.search)
-const url = params.get('key') ?? undefined
-
 ReactDOM.createRoot(app).render(<StrictMode>
-  <App url={url} />
+  <App />
 </StrictMode>)
